requirements.txt
pyqt5
pillow
numpy
matplotlib
piexif
opencv-python
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.8.0+cu126
torchvision==0.23.0+cu126
lpips